Create a SYM layer and replace LOWER with SYM+NUM.
This puts all the symbols under the left hand and puts a numpad under
the right hand without having to use a space as an extra layer key.

I moved the remaining space layer switcher (NAV) to the left space.
I type space with my right thumb, so I hit the right space.  Having the
space as a layer tap (presumably) made the keyboard wait until
I released the key to register a space.  This created a noticable delay
and felt weird to me.